# openapi.yaml
openapi: 3.1.0
info:
  title: Aku-Telhone
  description: Connectivity orchestration platform for the Aku ecosystem. Handles
    eSIM and physical SIM lifecycle management, customer onboarding, app-facing activation
    flows, hub policy integration, observability, and device attestation via Aku-IGHub.
  version: 0.1.0
paths:
...
  /api/v1/sims/inventory:
    post:
      tags:
      - physical-sims
      summary: Create a physical SIM inventory record
      operationId: create_inventory_record_api_v1_sims_inventory_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SIMInventoryRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhysicalSIMResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sims/{iccid}:
    get:
      tags:
      - physical-sims
      summary: Get a physical SIM record
      operationId: get_physical_sim_api_v1_sims__iccid__get
      parameters:
      - name: iccid
        in: path
        required: true
        schema:
          type: string
          title: Iccid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhysicalSIMResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - physical-sims
      summary: Deactivate a physical SIM
      operationId: deactivate_physical_sim_api_v1_sims__iccid__delete
      parameters:
      - name: iccid
        in: path
        required: true
        schema:
          type: string
          title: Iccid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhysicalSIMResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sims/{iccid}/assign:
    post:
      tags:
      - physical-sims
      summary: Assign a physical SIM to a customer/device
      operationId: assign_physical_sim_api_v1_sims__iccid__assign_post
      parameters:
      - name: iccid
        in: path
        required: true
        schema:
          type: string
          title: Iccid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SIMAssignRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhysicalSIMResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sims/{iccid}/activate:
    post:
      tags:
      - physical-sims
      summary: Activate an assigned physical SIM
      operationId: activate_physical_sim_api_v1_sims__iccid__activate_post
      parameters:
      - name: iccid
        in: path
        required: true
        schema:
          type: string
          title: Iccid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhysicalSIMResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sims/{iccid}/suspend:
    post:
      tags:
      - physical-sims
      summary: Suspend a physical SIM and linked subscription
      operationId: suspend_physical_sim_api_v1_sims__iccid__suspend_post
      parameters:
      - name: iccid
        in: path
        required: true
        schema:
          type: string
          title: Iccid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhysicalSIMResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sims/{iccid}/replace:
    post:
      tags:
      - physical-sims
      summary: Swap a physical SIM to a replacement SIM
      operationId: replace_physical_sim_api_v1_sims__iccid__replace_post
      parameters:
      - name: iccid
        in: path
        required: true
        schema:
          type: string
          title: Iccid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SIMReplaceRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhysicalSIMResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/app/sessions:
    post:
      tags:
      - app
      summary: Create an Aku-Telhone app session
      operationId: create_app_session_api_v1_app_sessions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppSessionRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppSessionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/app/customers/{customer_id}/dashboard:
    get:
      tags:
      - app
      summary: Get customer dashboard data for the Aku-Telhone app
      operationId: get_dashboard_api_v1_app_customers__customer_id__dashboard_get
      parameters:
      - name: customer_id
        in: path
        required: true
        schema:
          type: string
          title: Customer Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/app/subscriptions/{subscription_id}/activation:
    get:
      tags:
      - app
      summary: Poll activation progress and next steps for a line
      operationId: get_activation_status_api_v1_app_subscriptions__subscription_id__activation_get
      parameters:
      - name: subscription_id
        in: path
        required: true
        schema:
          type: string
          title: Subscription Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivationStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ActivationStatusResponse:
      properties:
        subscription:
          $ref: '#/components/schemas/SubscriptionResponse'
        ready_for_use:
          type: boolean
          title: Ready For Use
        next_step:
          type: string
          title: Next Step
      type: object
      required:
      - subscription
      - ready_for_use
      - next_step
      title: ActivationStatusResponse
    AppSessionRequest:
      properties:
        customer_id:
          type: string
          title: Customer Id
        channel:
          type: string
          title: Channel
          description: 'Client channel: mobile | web'
          default: mobile
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
      type: object
      required:
      - customer_id
      title: AppSessionRequest
    AppSessionResponse:
      properties:
        session_id:
          type: string
          title: Session Id
        customer_id:
          type: string
          title: Customer Id
        channel:
          type: string
          title: Channel
        issued_at:
          type: string
          format: date-time
          title: Issued At
        expires_at:
          type: string
          format: date-time
          title: Expires At
      type: object
      required:
      - session_id
      - customer_id
      - channel
      - expires_at
      title: AppSessionResponse
    BundleInfo:
      properties:
        bundle_id:
          type: string
          title: Bundle Id
        name:
          type: string
          title: Name
        category:
          type: string
          title: Category
          default: education
        active:
          type: boolean
          title: Active
          default: true
      type: object
      required:
      - bundle_id
      - name
      title: BundleInfo
    ComplianceStatus:
      type: string
      enum:
      - PENDING_REVIEW
      - APPROVED
      - RESTRICTED
      title: ComplianceStatus
...
    CustomerProfileResponse:
      properties:
        customer_id:
          type: string
          title: Customer Id
        full_name:
          type: string
          title: Full Name
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        phone_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone Number
        role:
          $ref: '#/components/schemas/UserRole'
        kyc_status:
          $ref: '#/components/schemas/KYCStatus'
        compliance_status:
          $ref: '#/components/schemas/ComplianceStatus'
        nin_reference:
          anyOf:
          - type: string
          - type: 'null'
          title: Nin Reference
        allowed_bundles:
          items:
            type: string
          type: array
          title: Allowed Bundles
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - customer_id
      - full_name
      - role
      - kyc_status
      - compliance_status
      title: CustomerProfileResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    KYCStatus:
      type: string
      enum:
      - PENDING
      - VERIFIED
      - REJECTED
      title: KYCStatus
    LineStatusResponse:
      properties:
        customer:
          $ref: '#/components/schemas/CustomerProfileResponse'
        subscriptions:
          items:
            $ref: '#/components/schemas/SubscriptionResponse'
          type: array
          title: Subscriptions
        available_bundles:
          items:
            $ref: '#/components/schemas/BundleInfo'
          type: array
          title: Available Bundles
        recent_audit_events:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Recent Audit Events
      type: object
      required:
      - customer
      title: LineStatusResponse
...
    NetworkTechnology:
      type: string
      enum:
      - LTE
      - LTE-M
      - NB-IoT
      - 5G
      title: NetworkTechnology
...
    PhysicalSIMResponse:
      properties:
        iccid:
          type: string
          title: Iccid
        serial_number:
          type: string
          title: Serial Number
        msisdn:
          anyOf:
          - type: string
          - type: 'null'
          title: Msisdn
        status:
          $ref: '#/components/schemas/PhysicalSIMStatus'
        plan_id:
          type: string
          title: Plan Id
        preferred_network:
          $ref: '#/components/schemas/NetworkTechnology'
        customer_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Customer Id
        subscription_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Subscription Id
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
        edge_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Edge Id
        activated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Activated At
        deactivated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deactivated At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - iccid
      - serial_number
      - status
      - plan_id
      - preferred_network
      title: PhysicalSIMResponse
    PhysicalSIMStatus:
      type: string
      enum:
      - INVENTORIED
      - ASSIGNED
      - ACTIVE
      - SUSPENDED
      - REPLACED
      - DEACTIVATED
      title: PhysicalSIMStatus
...
    SIMAssignRequest:
      properties:
        customer_id:
          type: string
          title: Customer Id
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
        activate_immediately:
          type: boolean
          title: Activate Immediately
          default: false
        allowed_bundles:
          items:
            type: string
          type: array
          title: Allowed Bundles
        fulfillment_reference:
          anyOf:
          - type: string
          - type: 'null'
          title: Fulfillment Reference
      type: object
      required:
      - customer_id
      title: SIMAssignRequest
    SIMInventoryRequest:
      properties:
        iccid:
          type: string
          minLength: 10
          title: Iccid
          description: Physical SIM ICCID
        serial_number:
          type: string
          title: Serial Number
          description: Printed SIM serial number
        msisdn:
          anyOf:
          - type: string
          - type: 'null'
          title: Msisdn
          description: Assigned Nigerian phone number when available
        plan_id:
          type: string
          title: Plan Id
          description: Default plan linked to the SIM inventory item
        preferred_network:
          $ref: '#/components/schemas/NetworkTechnology'
          default: LTE
        edge_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Edge Id
          description: Preferred Edge Hub for distribution
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
      type: object
      required:
      - iccid
      - serial_number
      - plan_id
      title: SIMInventoryRequest
    SIMReplaceRequest:
      properties:
        replacement_iccid:
          type: string
          title: Replacement Iccid
          description: New physical SIM ICCID
        reason:
          type: string
          maxLength: 512
          minLength: 3
          title: Reason
      type: object
      required:
      - replacement_iccid
      - reason
      title: SIMReplaceRequest
    SimFormFactor:
      type: string
      enum:
      - ESIM
      - PHYSICAL
      title: SimFormFactor
    SubscriptionResponse:
      properties:
        subscription_id:
          type: string
          title: Subscription Id
        customer_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Customer Id
        form_factor:
          $ref: '#/components/schemas/SimFormFactor'
        sim_identifier:
          type: string
          title: Sim Identifier
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
        status:
          $ref: '#/components/schemas/SubscriptionStatus'
        plan_id:
          type: string
          title: Plan Id
        preferred_network:
          $ref: '#/components/schemas/NetworkTechnology'
        activation_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Activation Code
        qr_code_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Qr Code Url
        allowed_bundles:
          items:
            type: string
          type: array
          title: Allowed Bundles
        activated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Activated At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - subscription_id
      - form_factor
      - sim_identifier
      - status
      - plan_id
      - preferred_network
      title: SubscriptionResponse
    SubscriptionStatus:
      type: string
      enum:
      - PENDING
      - READY
      - ACTIVE
      - SUSPENDED
      - SWITCHING
      - DEACTIVATED
      title: SubscriptionStatus
    UserRole:
      type: string
      enum:
      - STUDENT
      - TEACHER
      - STAFF
      - ADMIN
      title: UserRole
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError